Resolve #134: Allow fractional or 0 cpu jobs to be added. Fix #126: die when adding jobs with <0 cpus.

+++ cloud/cloud_test.go
@@ -212,11 +212,11 @@ func TestOpenStack(t *testing.T) {
 					n := server.HasSpaceFor(1, 0, 0)
 					So(n, ShouldEqual, flavor.Cores)
 
-					server.Allocate(flavor.Cores, 100, 0)
+					server.Allocate(float64(flavor.Cores), 100, 0)
 					n = server.HasSpaceFor(1, 0, 0)
 					So(n, ShouldEqual, 0)
 
-					server.Release(flavor.Cores, 100, 0)
+					server.Release(float64(flavor.Cores), 100, 0)
 					n = server.HasSpaceFor(1, 0, 0)
 					So(n, ShouldEqual, flavor.Cores)
